pyproject.toml
[project]
name = "preprocessing"
version = "0.1.0"
description = "PDF to RAG preprocessing pipeline with OCR, chunking, and Chroma vectorstore"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Codex"}]
license = {text = "MIT"}
dependencies = [
  "langchain>=0.2.14",
  "langchain-community>=0.2.12",
  "langchain-text-splitters>=0.2.2",
  "langchain-chroma>=0.1.4",
  "langchain-anthropic>=0.1.19",
  "sentence-transformers>=2.7.0",
  "chromadb>=0.5.7",
  "pypdf>=4.2.0",
  "pdfplumber>=0.11.0",
  "pymupdf>=1.24.0",
  "pdfminer.six>=20231228",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
  "tqdm>=4.66.0",
  "numpy>=1.26.0",
  "Pillow>=10.3.0",
  "pdf2image>=1.17.0",
  "pytesseract>=0.3.10",
  "easyocr>=1.7.1",
]

[project.optional-dependencies]
extras = [
  "unstructured[pdf]>=0.14.7",
  "nltk>=3.9",
  "spacy>=3.7",
  "beautifulsoup4>=4.12",
  "lxml>=5.2",
  "langchain-experimental>=0.0.64",
  "pytest>=8.0.0",
]

[project.scripts]
preprocess = "preprocessing.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
